Refacto '/api/locale' enpoint
