+++ src/pages/api/locale/index.ts
@@ -6,6 +6,7 @@ import {
   errorMiddleware,
   recoveryMiddleware,
   rateLimitingMiddleware,
+  Middleware,
 } from "@fwl/web/dist/middlewares";
 import { NextApiRequest, NextApiResponse } from "next";
 
@@ -22,7 +23,50 @@ import { ERRORS_DATA, webErrorFactory } from "@src/errors/web-errors";
 import { sentryMiddleware } from "@src/middlewares/sentry-middleware";
 import { getDBUserFromSub } from "@src/queries/get-db-user-from-sub";
 
-const handler: Handler = (request, response): Promise<void> => {
+const getHandler: Handler = (request, response): Promise<void> => {
+  const webErrors = {
+    databaseUnreachable: ERRORS_DATA.DATABASE_UNREACHABLE,
+  };
+
+  return getTracer().span("get-locale handler", async (span) => {
+    const userCookie = await getServerSideCookies<UserCookie>(request, {
+      cookieName: "user-cookie",
+      isCookieSealed: true,
+      cookieSalt: configVariables.cookieSalt,
+    });
+
+    if (!userCookie) {
+      span.setDisclosedAttribute("user cookie found", false);
+      throw new NoUserCookieFoundError();
+    }
+    span.setDisclosedAttribute("user cookie found", true);
+
+    const currentDBUser = await getDBUserFromSub(userCookie.sub).catch(
+      (error) => {
+        span.setDisclosedAttribute("user locale retrieved", false);
+        throw webErrorFactory({
+          ...webErrors.databaseUnreachable,
+          parentError: error,
+        });
+      },
+    );
+
+    if (!currentDBUser) {
+      span.setDisclosedAttribute("user locale value", null);
+      response.statusCode = HttpStatus.TEMPORARY_REDIRECT;
+      response.setHeader("location", "/");
+      response.end();
+      return;
+    }
+
+    span.setDisclosedAttribute("user locale retrieved", currentDBUser.locale);
+    response.statusCode = HttpStatus.OK;
+    response.json({ locale: currentDBUser.locale });
+    return;
+  });
+};
+
+const postHandler: Handler = (request, response): Promise<void> => {
   const webErrors = {
     badRequest: ERRORS_DATA.BAD_REQUEST,
     databaseUnreachable: ERRORS_DATA.DATABASE_UNREACHABLE,
@@ -67,19 +111,16 @@ const handler: Handler = (request, response): Promise<void> => {
   });
 };
 
-const wrappedHandler = wrapMiddlewares(
-  [
-    tracingMiddleware(getTracer()),
-    rateLimitingMiddleware(getTracer(), logger, rateLimitingConfig),
-    recoveryMiddleware(getTracer()),
-    sentryMiddleware(getTracer()),
-    errorMiddleware(getTracer()),
-    loggingMiddleware(getTracer(), logger),
-  ],
-  handler,
-  "/api/locale/set-locale",
-);
+const middlewares: Middleware<NextApiRequest, NextApiResponse>[] = [
+  tracingMiddleware(getTracer()),
+  rateLimitingMiddleware(getTracer(), logger, rateLimitingConfig),
+  recoveryMiddleware(getTracer()),
+  sentryMiddleware(getTracer()),
+  errorMiddleware(getTracer()),
+  loggingMiddleware(getTracer(), logger),
+];
 
 export default new Endpoint<NextApiRequest, NextApiResponse>()
-  .post(wrappedHandler)
+  .get(wrapMiddlewares(middlewares, getHandler, "GET /api/locale"))
+  .post(wrapMiddlewares(middlewares, postHandler, "POST /api/locale"))
   .getHandler();
